+++ Pfhorge Source/View and Controller/Others/PhColorListControllerDrawer.h
@@ -71,7 +71,7 @@
 - (void)updateInterface:(LEMapDraw *)theView;
 
 // ***********************Accsessor Methods ***********************
-- (int)getSelection;
+- (NSInteger)getSelection;
 - (NSNumber *)getSelectedNumber;
 
 - (void)setSelectionToNumber:(NSNumber *)theNumberToSelect;
